Represent exception type presenter as interface
